+++ spec/dummy/client/entries/app-kitten.js
@@ -22,6 +22,7 @@ import {
   LinkedinButtonIcon,
   TwitterButtonIcon,
   YoutubeButtonIcon,
+  PinterestButtonIcon,
 } from 'kitten/components/buttons/social-button-icon'
 import {
   KarlFacebookButtonIconWords,
@@ -42,6 +43,7 @@ import { KarlPaymentButtons } from 'kitten/karl/buttons/payment-button'
 import { KarlButtonIconFilter } from 'kitten/karl/buttons/button-filter'
 import { KarlButtonImageWithText } from 'kitten/karl/buttons/button-image-with-text'
 import { KarlGarbageButton } from 'kitten/karl/buttons/garbage-button'
+import { KarlStarButton } from 'kitten/karl/buttons/star-button'
 
 // Cards
 import { ArticleCard } from 'kitten/components/cards/article-card'
@@ -343,6 +345,7 @@ ReactOnRails.register({
   InstagramButtonIcon,
   LinkedinButtonIcon,
   YoutubeButtonIcon,
+  PinterestButtonIcon,
   TagButton,
   TwitterButtonIcon,
   KarlButtonAndText,
@@ -357,6 +360,7 @@ ReactOnRails.register({
   KarlButtonIconFilter,
   KarlButtonImageWithText,
   KarlGarbageButton,
+  KarlStarButton,
 
   // Colors
   KarlColors,